Revert "🔧 mobile gallery - refactor to remove eslint error (#72)"
This reverts commit 4e2a7e72c752f1f6ffc3f9d779d4e33eec3123e6.
